🎨 togglebutton 코드 개별 파일로 관리
